Altered SQL queries to now not only require LTV but also a mortgage term of same or fewer years than one offered by product. Prevented duplicate quotes from being saved to the quote table.

+++ saveasquote.php
@@ -13,17 +13,35 @@ if(isset($_POST['savequote']))
     $fee = $_POST['fee'];
     $total = validate($_POST['total']);
 
-    $insertquote = "INSERT INTO quotes 
-    (`user_id`, `mortgage_type`, `monthly_payment`, `interest_rate`, `product_fee`, `total_payable`) 
-    VALUES ('$userid', '$type', '$monthly', '$interest', '$fee', '$total')";
+    $check = "SELECT user_id, mortgage_type, monthly_payment, interest_rate, total_payable FROM quotes 
+    WHERE user_id = '$userid' AND mortgage_type = '$type' AND monthly_payment = '$monthly' 
+    AND interest_rate ='$interest' AND total_payable = '$total'";
     
-    mysqli_query($mysqli, $insertquote);
-    mysqli_close($mysqli);
+    $result = mysqli_query($mysqli, $check);
 
-    echo '<script type="text/javascript">'; 
-    echo 'alert("Quote Saved");';
-    echo 'window.location.href = "productsearch.php";';
-    echo '</script>';
+    $returnedresults = mysqli_num_rows($result);
+
+    if($returnedresults==0) // If there are no results, meaning the record doesn't already exist.
+    {
+        $insertquote = "INSERT INTO quotes 
+        (`user_id`, `mortgage_type`, `monthly_payment`, `interest_rate`, `product_fee`, `total_payable`) 
+        VALUES ('$userid', '$type', '$monthly', '$interest', '$fee', '$total')";
+        
+        mysqli_query($mysqli, $insertquote);
+        mysqli_close($mysqli);
+    
+        echo '<script type="text/javascript">'; 
+        echo 'alert("Quote Saved");';
+        echo 'window.location.href = "productsearch.php";';
+        echo '</script>';
+    }
+    else //This would be a duplicate record, already saved to quotes. Don't save it.
+    {
+        echo '<script type="text/javascript">'; 
+        echo 'alert("Quote is already saved. Cannot save again.");';
+        echo 'window.location.href = "productsearch.php";';
+        echo '</script>';
+    }
 }
 else 
 {
